Added post for image method and properties data. Added photo resizer in UserService and fix deletePhoto. Remove findAll in post

+++ src/main/java/com/kochetkova/service/impl/UserServiceImpl.java
@@ -2,24 +2,32 @@ package com.kochetkova.service.impl;
 
 import com.kochetkova.api.request.EditProfileRequest;
 import com.kochetkova.api.request.NewUserRequest;
+import com.kochetkova.api.request.ResetPasswordRequest;
 import com.kochetkova.api.response.ErrorResponse;
 import com.kochetkova.api.response.ResultErrorResponse;
 import com.kochetkova.api.response.UserResponse;
+import com.kochetkova.model.ModerationStatus;
 import com.kochetkova.model.User;
 import com.kochetkova.repository.UserRepository;
 import com.kochetkova.service.MailSender;
 import com.kochetkova.service.UserService;
+import org.apache.commons.io.FilenameUtils;
+import org.imgscalr.Scalr;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.util.FileCopyUtils;
 import org.springframework.web.multipart.MultipartFile;
 
+import javax.imageio.ImageIO;
+import java.awt.image.BufferedImage;
 import java.io.*;
 import java.time.LocalDateTime;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.UUID;
+import java.util.stream.Collector;
+import java.util.stream.Collectors;
 
 @Service
 public class UserServiceImpl implements UserService {
@@ -30,8 +38,16 @@ public class UserServiceImpl implements UserService {
     private final String separator = File.separator;
     private static Map<String, Integer> sessions = new HashMap<>();
 
-    @Value("${photo.path}")
+    @Value("${photo.avatarPath}")
     private String photoPath;
+
+
+    @Value("${photo.width}")
+    private int photoWidth;
+
+
+    @Value("${photo.height}")
+    private int photoHeight;
 
     @Value("${password.length.min}")
     private int passwordLengthMin;
@@ -164,7 +180,17 @@ public class UserServiceImpl implements UserService {
         return errorBuilder.build();
     }
 
-    //сохранить данные пользователя для обновления профиля запрос с фото
+    /**
+     * сохранить данные пользователя для обновления профиля запрос с фото
+     *
+     * @param user        - пользователь
+     * @param name        - имя
+     * @param email       - почта
+     * @param password    - пароль
+     * @param photo       - фотография
+     * @param removePhoto - удалить фото при 1, сменить фото - 0
+     * @return User - пользователь с изменными данными
+     */
     @Override
     public User saveEditProfile(User user, String name, String email, String password, MultipartFile photo, Integer removePhoto) {
         if (!user.getName().equals(name)) {
@@ -186,7 +212,8 @@ public class UserServiceImpl implements UserService {
         }
         if (removePhoto != null && removePhoto == 1) {
             //delete photo
-            user.setPhoto(deletePhoto(user));
+            deletePhoto(user);
+            user.setPhoto("");
         }
 
         saveUser(user);
@@ -219,29 +246,33 @@ public class UserServiceImpl implements UserService {
         if (sessions.get(sessionId) == null) {
             return null;
         }
-       return findUserById(sessions.get(sessionId));
+        return findUserById(sessions.get(sessionId));
     }
 
     /**
-     *  cохранение новой фотографии в папке и установка ссылки у пользователя
+     * cохранение новой фотографии в папке и установка ссылки у пользователя
+     *
+     * @param user  - пользователь
+     * @param photo - фотография
+     * @return строка с путем к фотографии
      */
     @Override
     public String savePhoto(User user, MultipartFile photo) {
-        //todo
-        //изменение размера фото
-        //при загрузке файла изображения фотографии пользователя, необходимо выполнять обрезку и изменение размера фотографии до 36х36 пикселей.
+        deletePhoto(user);
 
         if (!photo.isEmpty()) {
             int id = user.getId();
 
-            int numberOfFolder = id / 100;
-            String fullPath = photoPath + separator + numberOfFolder + separator + id + "_" + photo.getOriginalFilename();
+            int numberOfFolder = id / 100; //разбито на папки по 100 пользователей
+            String fullPath = photoPath + separator + numberOfFolder + separator + id + "." + FilenameUtils.getExtension(photo.getOriginalFilename());
             File file = new File(fullPath);
             if (!file.getParentFile().exists()) {
                 file.getParentFile().mkdirs();
             }
             try {
-                FileCopyUtils.copy(photo.getBytes(), file);
+                BufferedImage resizePhoto = resizePhoto(photo);
+
+                ImageIO.write(resizePhoto, FilenameUtils.getExtension(photo.getOriginalFilename()), file);
             } catch (IOException e) {
                 e.printStackTrace();
             }
@@ -252,16 +283,39 @@ public class UserServiceImpl implements UserService {
     }
 
     /**
-     * удаление фотографии из папки и ссылки у пользователя
+     * Изменение размера фотографии под параметры (ширина и высота), заданные в настройках приложения
+     *
+     * @param photo - исходная фотография
+     * @return BufferedImage - параметризированная фотография
+     * @throws IOException
+     */
+    private BufferedImage resizePhoto(MultipartFile photo) throws IOException {
+        BufferedImage image = ImageIO.read(photo.getInputStream());
+
+        return Scalr.resize(image,
+                Scalr.Method.QUALITY,
+                Scalr.Mode.FIT_EXACT,
+                photoWidth,
+                photoHeight,
+                Scalr.OP_ANTIALIAS);
+    }
+
+    /**
+     * удаление фотографии из папки для аватара пользователя
+     *
+     * @param user - пользователь
+     * @return удалено - true или нет - false
      */
     @Override
-    public String deletePhoto(User user) {
+    public boolean deletePhoto(User user) {
+
         String fullPath = user.getPhoto();
+        if (fullPath.equalsIgnoreCase("")) {
+            return true;
+        }
         File file = new File(fullPath.substring(1).replace("/", "\\"));
-        file.delete();
-        user.setPhoto("");
-        userRepository.save(user);
-        return null;
+
+        return file.delete();
     }
 
     /**
@@ -281,6 +335,7 @@ public class UserServiceImpl implements UserService {
 
     /**
      * формирование UserResponse на основе User
+     *
      * @param user - информация, полученная из БД, о пользователе
      * @param mode - режим полноты выдачи информации:
      *             1 - только id и name;
@@ -312,7 +367,8 @@ public class UserServiceImpl implements UserService {
             userResponseBuilder.moderation(true);
             userResponseBuilder.setting(true);
         }
-        userResponseBuilder.moderationCount(userInfo.getModerationPosts().size());
+        userResponseBuilder.moderationCount((int) userInfo.getModerationPosts().stream()
+                .filter(post -> post.getModerationStatus() == ModerationStatus.NEW && post.getIsActive() == 1).count());
         return userResponseBuilder.build();
     }
 
@@ -321,6 +377,7 @@ public class UserServiceImpl implements UserService {
      * POST /api/auth/restore
      * Авторизация: не требуется
      * Если пользователь найден, ему должно отправляться письмо со ссылкой на восстановление пароля следующего вида - /login/change-password/HASH.
+     *
      * @param email - e-mail пользователя
      * @return ResultErrorResponse "result": true или false
      */
@@ -329,18 +386,34 @@ public class UserServiceImpl implements UserService {
         ResultErrorResponse result = new ResultErrorResponse();
         User user = findUserByEmail(email);
         if (user != null) {
-            user.setCode(UUID.randomUUID().toString());
+            user.setCode(UUID.randomUUID().toString().replace("-", ""));
             saveUser(user);
             String message = String.format("Hello, dear %s!\n"
-                    + "Вы отправили запрос на восстановление пароля?\n"
-                    + "Кто-то (надеемся, что вы) попросил нас сбросить пароль для вашей учетной записи. Чтобы сделать это, щелкните по ссылке ниже:\n"
-                    + "http://localhost:8080/login/change-password/%s\n"
-                    + "Если вы не запрашивали сброс пароля, игнорируйте это сообщение!",
+                            + "Вы отправили запрос на восстановление пароля?\n"
+                            + "Кто-то (надеемся, что вы) попросил нас сбросить пароль для вашей учетной записи. Чтобы сделать это, щелкните по ссылке ниже:\n"
+                            + "http://localhost:8080/login/change-password/%s\n"
+                            + "Если вы не запрашивали сброс пароля, игнорируйте это сообщение!",
                     user.getName(), user.getCode());
             mailSender.send(user.getEmail(), "Восстановление пароля", message);
 
             result.setResult(true);
         }
         return result;
+    }
+
+    /**
+     * Изменение пароля
+     * POST /api/auth/password
+     * Авторизация: не требуется
+     *
+     * @param resetPasswordRequest - код восстановления пароля и коды капчи
+     * @return ResultErrorResponse
+     * В случае, если все данные отправлены верно: "result": true
+     * В случае ошибок: "result": false + "errors":
+     */
+    @Override
+    public ResultErrorResponse setNewPassword(ResetPasswordRequest resetPasswordRequest) {
+
+        return null;
     }
 }
